# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/me": {
      "get": {
        "operationId": "getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "status",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "$ref": "#/components/schemas/MeProfile"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Lỗi, error.code thuộc: UNAUTHORIZED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                },
                "examples": {
                  "UNAUTHORIZED": {
                    "summary": "UNAUTHORIZED: Bạn cần đăng nhập để tiếp tục",
                    "value": {
                      "success": false,
                      "status": 401,
                      "error": {
                        "code": "UNAUTHORIZED",
                        "message": "Bạn cần đăng nhập để tiếp tục"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Lỗi, error.code thuộc: NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                },
                "examples": {
                  "NOT_FOUND": {
                    "summary": "NOT_FOUND: Tài khoản không còn tồn tại",
                    "value": {
                      "success": false,
                      "status": 404,
                      "error": {
                        "code": "NOT_FOUND",
                        "message": "Tài khoản không còn tồn tại"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Lỗi, error.code thuộc: INTERNAL_SERVER_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                },
                "examples": {
                  "INTERNAL_SERVER_ERROR": {
                    "summary": "INTERNAL_SERVER_ERROR: Đã có lỗi xảy ra, vui lòng thử lại sau",
                    "value": {
                      "success": false,
                      "status": 500,
                      "error": {
                        "code": "INTERNAL_SERVER_ERROR",
                        "message": "Đã có lỗi xảy ra, vui lòng thử lại sau"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy thông tin hồ sơ cá nhân",
        "tags": [
          "Me"
        ]
      },
      "patch": {
        "description": "Trường nào không gửi thì giữ nguyên; targetScore chỉ dành cho học viên.",
        "operationId": "updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "status",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "$ref": "#/components/schemas/MeProfile"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Lỗi, error.code thuộc: VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                },
                "examples": {
                  "VALIDATION_ERROR": {
                    "summary": "VALIDATION_ERROR: Dữ liệu gửi lên không hợp lệ",
                    "value": {
                      "success": false,
                      "status": 400,
                      "error": {
                        "code": "VALIDATION_ERROR",
                        "message": "Dữ liệu gửi lên không hợp lệ",
                        "details": [
                          {
                            "field": "fullName",
                            "message": "fullName không hợp lệ"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Lỗi, error.code thuộc: UNAUTHORIZED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                },
                "examples": {
                  "UNAUTHORIZED": {
                    "summary": "UNAUTHORIZED: Bạn cần đăng nhập để tiếp tục",
                    "value": {
                      "success": false,
                      "status": 401,
                      "error": {
                        "code": "UNAUTHORIZED",
                        "message": "Bạn cần đăng nhập để tiếp tục"
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Lỗi, error.code thuộc: FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                },
                "examples": {
                  "FORBIDDEN": {
                    "summary": "FORBIDDEN: Chỉ học viên mới đặt được điểm mục tiêu",
                    "value": {
                      "success": false,
                      "status": 403,
                      "error": {
                        "code": "FORBIDDEN",
                        "message": "Chỉ học viên mới đặt được điểm mục tiêu"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Lỗi, error.code thuộc: NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                },
                "examples": {
                  "NOT_FOUND": {
                    "summary": "NOT_FOUND: Tài khoản không còn tồn tại",
                    "value": {
                      "success": false,
                      "status": 404,
                      "error": {
                        "code": "NOT_FOUND",
                        "message": "Tài khoản không còn tồn tại"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Lỗi, error.code thuộc: INTERNAL_SERVER_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                },
                "examples": {
                  "INTERNAL_SERVER_ERROR": {
                    "summary": "INTERNAL_SERVER_ERROR: Đã có lỗi xảy ra, vui lòng thử lại sau",
                    "value": {
                      "success": false,
                      "status": 500,
                      "error": {
                        "code": "INTERNAL_SERVER_ERROR",
                        "message": "Đã có lỗi xảy ra, vui lòng thử lại sau"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cập nhật hồ sơ cá nhân",
        "tags": [
          "Me"
        ]
      }
    },
    "/api/v1/me/login-history": {
      "get": {
        "operationId": "getLoginHistory",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "status",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/LoginHistoryItem"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PaginationMetaDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Lỗi, error.code thuộc: VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                },
                "examples": {
                  "VALIDATION_ERROR": {
                    "summary": "VALIDATION_ERROR: Dữ liệu gửi lên không hợp lệ",
                    "value": {
                      "success": false,
                      "status": 400,
                      "error": {
                        "code": "VALIDATION_ERROR",
                        "message": "Dữ liệu gửi lên không hợp lệ",
                        "details": [
                          {
                            "field": "page",
                            "message": "page không hợp lệ"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Lỗi, error.code thuộc: UNAUTHORIZED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                },
                "examples": {
                  "UNAUTHORIZED": {
                    "summary": "UNAUTHORIZED: Bạn cần đăng nhập để tiếp tục",
                    "value": {
                      "success": false,
                      "status": 401,
                      "error": {
                        "code": "UNAUTHORIZED",
                        "message": "Bạn cần đăng nhập để tiếp tục"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Lỗi, error.code thuộc: INTERNAL_SERVER_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                },
                "examples": {
                  "INTERNAL_SERVER_ERROR": {
                    "summary": "INTERNAL_SERVER_ERROR: Đã có lỗi xảy ra, vui lòng thử lại sau",
                    "value": {
                      "success": false,
                      "status": 500,
                      "error": {
                        "code": "INTERNAL_SERVER_ERROR",
                        "message": "Đã có lỗi xảy ra, vui lòng thử lại sau"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Xem lịch sử đăng nhập của chính mình, mới nhất trước",
        "tags": [
          "Me"
        ]
      }
    },
    "/api/v1/me/password": {
      "patch": {
        "description": "Thành công thì đăng xuất mọi thiết bị; FE gọi lại POST /auth/login bằng mật khẩu mới.",
        "operationId": "changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "status",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "type": "object",
                      "nullable": true,
                      "example": null
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Lỗi, error.code thuộc: AUTH_CURRENT_PASSWORD_INCORRECT, BAD_REQUEST, VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                },
                "examples": {
                  "AUTH_CURRENT_PASSWORD_INCORRECT": {
                    "summary": "AUTH_CURRENT_PASSWORD_INCORRECT: Mật khẩu hiện tại không đúng",
                    "value": {
                      "success": false,
                      "status": 400,
                      "error": {
                        "code": "AUTH_CURRENT_PASSWORD_INCORRECT",
                        "message": "Mật khẩu hiện tại không đúng"
                      }
                    }
                  },
                  "BAD_REQUEST": {
                    "summary": "BAD_REQUEST: Mật khẩu mới phải khác mật khẩu hiện tại",
                    "value": {
                      "success": false,
                      "status": 400,
                      "error": {
                        "code": "BAD_REQUEST",
                        "message": "Mật khẩu mới phải khác mật khẩu hiện tại"
                      }
                    }
                  },
                  "VALIDATION_ERROR": {
                    "summary": "VALIDATION_ERROR: Dữ liệu gửi lên không hợp lệ",
                    "value": {
                      "success": false,
                      "status": 400,
                      "error": {
                        "code": "VALIDATION_ERROR",
                        "message": "Dữ liệu gửi lên không hợp lệ",
                        "details": [
                          {
                            "field": "currentPassword",
                            "message": "currentPassword không hợp lệ"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Lỗi, error.code thuộc: UNAUTHORIZED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                },
                "examples": {
                  "UNAUTHORIZED": {
                    "summary": "UNAUTHORIZED: Bạn cần đăng nhập để tiếp tục",
                    "value": {
                      "success": false,
                      "status": 401,
                      "error": {
                        "code": "UNAUTHORIZED",
                        "message": "Bạn cần đăng nhập để tiếp tục"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Lỗi, error.code thuộc: NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                },
                "examples": {
                  "NOT_FOUND": {
                    "summary": "NOT_FOUND: Tài khoản không còn tồn tại",
                    "value": {
                      "success": false,
                      "status": 404,
                      "error": {
                        "code": "NOT_FOUND",
                        "message": "Tài khoản không còn tồn tại"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Lỗi, error.code thuộc: INTERNAL_SERVER_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                },
                "examples": {
                  "INTERNAL_SERVER_ERROR": {
                    "summary": "INTERNAL_SERVER_ERROR: Đã có lỗi xảy ra, vui lòng thử lại sau",
                    "value": {
                      "success": false,
                      "status": 500,
                      "error": {
                        "code": "INTERNAL_SERVER_ERROR",
                        "message": "Đã có lỗi xảy ra, vui lòng thử lại sau"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Đổi mật khẩu",
        "tags": [
          "Me"
        ]
      }
    }
  },
  "info": {
    "title": "Smart English Exam System API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token"
      }
    },
    "schemas": {
      "ErrorCode": {
        "type": "string",
        "enum": [
          "BAD_REQUEST",
          "VALIDATION_ERROR",
          "UNAUTHORIZED",
          "FORBIDDEN",
          "NOT_FOUND",
          "CONFLICT",
          "VERSION_CONFLICT",
          "GONE",
          "PAYLOAD_TOO_LARGE",
          "UNPROCESSABLE_ENTITY",
          "TOO_MANY_REQUESTS",
          "INTERNAL_SERVER_ERROR",
          "SERVICE_UNAVAILABLE",
          "AUTH_INVALID_CREDENTIALS",
          "AUTH_EMAIL_NOT_VERIFIED",
          "AUTH_ACCOUNT_PENDING_APPROVAL",
          "AUTH_ACCOUNT_LOCKED",
          "AUTH_REFRESH_TOKEN_INVALID",
          "AUTH_EMAIL_ALREADY_EXISTS",
          "AUTH_TOKEN_INVALID",
          "AUTH_TOO_MANY_LOGIN_ATTEMPTS",
          "AUTH_CURRENT_PASSWORD_INCORRECT"
        ]
      },
      "FieldErrorDto": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "example": "email",
            "description": "Trường lỗi, trường lồng nhau dạng address.city"
          },
          "message": {
            "type": "string",
            "example": "Email không hợp lệ"
          }
        },
        "required": [
          "field",
          "message"
        ]
      },
      "ErrorBodyDto": {
        "type": "object",
        "properties": {
          "code": {
            "example": "VALIDATION_ERROR",
            "allOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              }
            ]
          },
          "message": {
            "type": "string",
            "example": "Dữ liệu gửi lên không hợp lệ"
          },
          "details": {
            "description": "VALIDATION_ERROR: danh sách lỗi theo trường. Mã khác có thể chứa dữ liệu riêng, ví dụ { retryAfterSeconds }",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldErrorDto"
            }
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ErrorEnvelopeDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "status": {
            "type": "integer",
            "example": 400,
            "description": "Luôn trùng HTTP status code"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorBodyDto"
          }
        },
        "required": [
          "success",
          "status",
          "error"
        ]
      },
      "UserStatus": {
        "type": "string",
        "enum": [
          "PENDING_VERIFICATION",
          "PENDING_APPROVAL",
          "ACTIVE",
          "LOCKED"
        ]
      },
...
      "RoleCode": {
        "type": "string",
        "enum": [
          "STUDENT",
          "TEACHER",
          "ADMIN"
        ]
      },
...
      "Level": {
        "type": "string",
        "enum": [
          "BEGINNER",
          "INTERMEDIATE",
          "ADVANCED"
        ]
      },
      "StudentProfile": {
        "type": "object",
        "properties": {
          "currentLevel": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/Level"
              }
            ]
          },
          "currentScore": {
            "type": "integer",
            "nullable": true,
            "example": 550
          },
          "targetScore": {
            "type": "integer",
            "nullable": true,
            "example": 750
          }
        },
        "required": [
          "currentLevel",
          "currentScore",
          "targetScore"
        ]
      },
      "MeProfile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "example": "hocvien@gmail.com"
          },
          "fullName": {
            "type": "string",
            "example": "Nguyễn Văn A"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RoleCode"
              }
            ]
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserStatus"
              }
            ]
          },
          "mustChangePassword": {
            "type": "boolean"
          },
          "emailVerifiedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastLoginAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "student": {
            "nullable": true,
            "description": "null với giáo viên và admin",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/StudentProfile"
              }
            ]
          }
        },
        "required": [
          "id",
          "email",
          "fullName",
          "avatarUrl",
          "role",
          "status",
          "mustChangePassword",
          "emailVerifiedAt",
          "lastLoginAt",
          "createdAt",
          "updatedAt",
          "student"
        ]
      },
      "LoginHistoryItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "success": {
            "type": "boolean"
          },
          "failureReason": {
            "type": "string",
            "nullable": true,
            "example": "INVALID_PASSWORD",
            "description": "Mã lý do khi thất bại, ví dụ INVALID_PASSWORD, TOO_MANY_ATTEMPTS; null khi thành công"
          },
          "ipAddress": {
            "type": "string",
            "nullable": true,
            "example": "127.0.0.1"
          },
          "userAgent": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "success",
          "failureReason",
          "ipAddress",
          "userAgent",
          "createdAt"
        ]
      },
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 20
          },
          "total": {
            "type": "integer",
            "example": 42
          },
          "totalPages": {
            "type": "integer",
            "example": 3,
            "description": "0 khi không có bản ghi nào"
          }
        },
        "required": [
          "page",
          "limit",
          "total",
          "totalPages"
        ]
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "example": "Nguyễn Văn A",
            "minLength": 2,
            "maxLength": 100
          },
          "targetScore": {
            "type": "number",
            "example": 750,
            "minimum": 10,
            "maximum": 990,
            "description": "Điểm TOEIC Listening + Reading mục tiêu, bước 5 điểm (chỉ học viên)"
          }
        }
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "example": "MatKhau123"
          },
          "newPassword": {
            "type": "string",
            "example": "MatKhauMoi123",
            "minLength": 8,
            "maxLength": 72,
            "description": "Mật khẩu dài 8-72 ký tự, gồm ít nhất một chữ hoa, một chữ thường và một chữ số"
          },
          "confirmPassword": {
            "type": "string",
            "example": "MatKhauMoi123"
          }
        },
        "required": [
          "currentPassword",
          "newPassword",
          "confirmPassword"
        ]
      }
    }
  }
}